requirements.txt
python-ranges==0.1.3
pandas==1.0.1
word2number==1.1
matplotlib==3.1.3
numpy==1.18.1
spacy==2.3.2
pytest==5.3.5
seaborn==0.10.0
scipy==1.4.1
#python -m spacy download en_core_web_sm
#python -m spacy download en_core_web_lg